clean up code and break it into smaller files
