+++ src/screens/points/ConsoleSheet.tsx
@@ -1,44 +1,16 @@
 /* eslint-disable no-nested-ternary */
 import React, {
-  createContext,
+  forwardRef,
   useCallback,
-  useContext,
   useEffect,
-  useMemo,
+  useImperativeHandle,
   useRef,
   useState,
 } from 'react';
-import { StyleSheet, Text as RNText, View, Linking } from 'react-native';
-import ReactNativeHapticFeedback from 'react-native-haptic-feedback';
-import Animated, {
-  Easing,
-  cancelAnimation,
-  runOnJS,
-  useAnimatedReaction,
-  useSharedValue,
-  withRepeat,
-  withSequence,
-  withTiming,
-} from 'react-native-reanimated';
-import { AnimatePresence } from '@/components/animations/AnimatePresence';
-import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
-import {
-  Bleed,
-  Box,
-  Cover,
-  Inline,
-  Inset,
-  Stack,
-  Text,
-  globalColors,
-  useForegroundColor,
-} from '@/design-system';
-import { alignHorizontalToFlexAlign } from '@/design-system/layout/alignment';
-import { useAccountProfile, useDimensions } from '@/hooks';
-import { fonts } from '@/styles';
-import { useTheme } from '@/theme';
-import { safeAreaInsetValues } from '@/utils';
-import { HapticFeedbackType } from '@/utils/haptics';
+import { StyleSheet } from 'react-native';
+import Animated from 'react-native-reanimated';
+import { Box, Inset, Stack, globalColors } from '@/design-system';
+import { useAccountProfile } from '@/hooks';
 import { metadataPOSTClient } from '@/graphql';
 import { signPersonalMessage } from '@/model/wallet';
 import { RouteProp, useRoute } from '@react-navigation/native';
@@ -52,7 +24,6 @@ import { pointsQueryKey } from '@/resources/points';
 import { queryClient } from '@/react-query';
 import * as i18n from '@/languages';
 import { abbreviateNumber } from '@/helpers/utilities';
-import Routes from '@/navigation/routesNames';
 import {
   address as formatAddress,
   abbreviateEnsForDisplay,
@@ -60,9 +31,18 @@ import {
 import { usePointsTweetIntentQuery } from '@/resources/pointsTweetIntent/pointsTweetIntentQuery';
 import { useNavigation } from '@/navigation';
 
-const SCREEN_BOTTOM_INSET = safeAreaInsetValues.bottom + 20;
-const CHARACTER_WIDTH = 9.2725;
-const POINTS_TWEET_INTENT_ID = '3ttGOpIsbJg3aa00FScp3I';
+import Line from './components/Line';
+import {
+  POINTS_TWEET_INTENT_ID,
+  SCREEN_BOTTOM_INSET,
+  rainbowColors,
+  rainbowText,
+  textColors,
+} from './constants';
+import { TypingAnimation } from './contexts/AnimationContext';
+import Paragraph from './components/Paragraph';
+import { AnimatedText } from './components/AnimatedText';
+import LineBreak from './components/LineBreak';
 
 type ConsoleSheetParams = {
   ConsoleSheet: {
@@ -70,35 +50,21 @@ type ConsoleSheetParams = {
   };
 };
 
+type ClaimFlowRef = {
+  isCalculationComplete: boolean;
+  setIsCalculationComplete: React.Dispatch<React.SetStateAction<boolean>>;
+};
+
 export const ConsoleSheet = () => {
   const { params } = useRoute<RouteProp<ConsoleSheetParams, 'ConsoleSheet'>>();
   const referralCode = params?.referralCode;
 
+  const claimFlowRef = useRef<ClaimFlowRef>(null);
   const { navigate } = useNavigation();
-
-  const [didConfirmOwnership, setDidConfirmOwnership] = useState(false);
-  const [showSignInButton, setShowSignInButton] = useState(false);
-  const [animationPhase, setAnimationPhase] = useState(0);
-  const [isCalculationComplete, setIsCalculationComplete] = useState(false);
-  const [showShareButtons, setShowShareButtons] = useState(false);
-  const [tweetIntent, setTweetIntent] = useState<string>('');
-  const [didShare, setDidShare] = useState(false);
-  const [showDoneButton, setShowDoneButton] = useState(false);
 
   const [pointsProfile, setPointsProfile] = useState<OnboardPointsMutation>();
 
   const { accountAddress } = useAccountProfile();
-
-  useEffect(() => {
-    setDidConfirmOwnership(false);
-    setShowSignInButton(false);
-    setAnimationPhase(0);
-    setIsCalculationComplete(false);
-    setShowShareButtons(false);
-    setTweetIntent('');
-    setDidShare(false);
-    setShowDoneButton(false);
-  }, []);
 
   const signIn = useCallback(async () => {
     let points;
@@ -172,22 +138,15 @@ export const ConsoleSheet = () => {
             width={{ custom: 36 }}
           />
           <ClaimRetroactivePointsFlow
+            ref={claimFlowRef}
             pointsProfile={pointsProfile}
-            animationPhase={animationPhase}
-            setAnimationPhase={setAnimationPhase}
-            didConfirmOwnership={didConfirmOwnership}
-            setShowSignInButton={setShowSignInButton}
-            isCalculationComplete={isCalculationComplete}
-            setIsCalculationComplete={setIsCalculationComplete}
-            setShowShareButtons={setShowShareButtons}
-            setTweetIntent={setTweetIntent}
-            didShare={didShare}
-            setShowDoneButton={setShowDoneButton}
           />
         </Animated.View>
       </Box>
-      <AnimatePresence
-        condition={showSignInButton && !didConfirmOwnership}
+      {/* <AnimatePresence
+        condition={
+          claimFlowRef.current?.showSignInButton && !didConfirmOwnership
+        }
         duration={300}
       >
         <NeonButton
@@ -196,7 +155,9 @@ export const ConsoleSheet = () => {
         />
       </AnimatePresence>
       <AnimatePresence
-        condition={animationPhase === 1 && isCalculationComplete}
+        condition={
+          animationPhase === 1 && claimFlowRef.current?.isCalculationComplete
+        }
         duration={300}
       >
         <NeonButton
@@ -239,413 +200,425 @@ export const ConsoleSheet = () => {
           label={`􀖅 ${i18n.t(i18n.l.points.console.complete_onboarding)}`}
           onPress={() => navigate(Routes.POINTS_SCREEN)}
         />
-      </AnimatePresence>
+      </AnimatePresence> */}
     </Inset>
   );
 };
 
-const NeonButton = ({
-  color,
-  label,
-  onPress,
-}: {
-  color?: string;
-  label: string;
-  onPress?: () => void;
-}) => {
-  const { width: deviceWidth } = useDimensions();
-  const { colors } = useTheme();
-  const green = useForegroundColor('green');
-
-  return (
-    <Box
-      alignItems="center"
-      width="full"
-      justifyContent="center"
-      position="absolute"
-      bottom={{ custom: 16 }}
-    >
-      <ButtonPressAnimation
-        hapticType="impactHeavy"
-        onPress={onPress}
-        scaleTo={0.94}
-        style={styles.neonButtonWrapper}
-        transformOrigin="top"
-      >
-        <Animated.View
-          style={[
-            styles.neonButton,
-            {
-              alignItems: 'center',
-              justifyContent: 'center',
-              borderColor: color || green,
-              shadowColor: color || green,
-              width: deviceWidth - 64,
-            },
-          ]}
-        >
-          <Cover>
-            <Box
-              borderRadius={11}
-              height={{ custom: 46 }}
-              style={[
-                styles.neonButtonFill,
-                {
-                  backgroundColor: colors.alpha(color || green, 0.1),
-                },
-              ]}
-              width={{ custom: deviceWidth - 66 }}
-            />
-          </Cover>
-          <RNText
-            style={[
-              styles.neonButtonText,
-              {
-                textShadowColor: colors.alpha(color || green, 0.6),
-              },
-            ]}
-          >
-            <Text
-              align="center"
-              color={color ? { custom: color } : 'green'}
-              size="20pt"
-              weight="heavy"
-            >
-              {label}
-            </Text>
-          </RNText>
-        </Animated.View>
-      </ButtonPressAnimation>
-    </Box>
-  );
-};
-
-const ClaimRetroactivePointsFlow = ({
-  pointsProfile,
-  animationPhase,
-  setAnimationPhase,
-  didConfirmOwnership,
-  setShowSignInButton,
-  isCalculationComplete,
-  setIsCalculationComplete,
-  setShowShareButtons,
-  setTweetIntent,
-  didShare,
-  setShowDoneButton,
-}: {
-  pointsProfile?: OnboardPointsMutation;
-  animationPhase: number;
-  setAnimationPhase: React.Dispatch<React.SetStateAction<number>>;
-  didConfirmOwnership: boolean;
-  setShowSignInButton: React.Dispatch<React.SetStateAction<boolean>>;
-  isCalculationComplete: boolean;
-  setIsCalculationComplete: React.Dispatch<React.SetStateAction<boolean>>;
-  setShowShareButtons: React.Dispatch<React.SetStateAction<boolean>>;
-  setTweetIntent: React.Dispatch<React.SetStateAction<string>>;
-  didShare: boolean;
-  setShowDoneButton: React.Dispatch<React.SetStateAction<boolean>>;
-}) => {
-  const { accountENS, accountAddress } = useAccountProfile();
-  const [animationKey, setAnimationKey] = useState(0);
-
-  const accountName = (abbreviateEnsForDisplay(accountENS, 10) ||
-    formatAddress(accountAddress, 4, 5)) as string;
-
-  useEffect(() => {
-    setAnimationKey(prevKey => prevKey + 1);
-  }, []);
-
-  usePointsTweetIntentQuery(
+const ClaimRetroactivePointsFlow = forwardRef(
+  (
     {
-      id: POINTS_TWEET_INTENT_ID,
+      pointsProfile,
+    }: {
+      pointsProfile?: OnboardPointsMutation;
     },
-    {
-      enabled: animationPhase === 2,
-      onSuccess: data => {
-        if (!data.pointsTweetIntent) {
-          return;
-        }
+    ref
+  ) => {
+    const { accountENS, accountAddress } = useAccountProfile();
+    const [animationKey, setAnimationKey] = useState(0);
 
-        try {
-          const tweetIntent = data.pointsTweetIntent;
-          const pointsValue =
-            pointsProfile?.onboardPoints?.user.onboarding.earnings.total;
-          if (!pointsValue) {
-            // TODO: Fallback to some default msg
+    const [isCalculationComplete, setIsCalculationComplete] = useState(false);
+    const [didConfirmOwnership, setDidConfirmOwnership] = useState(false);
+    const [showSignInButton, setShowSignInButton] = useState(false);
+    const [animationPhase, setAnimationPhase] = useState(0);
+    const [showShareButtons, setShowShareButtons] = useState(false);
+    const [tweetIntent, setTweetIntent] = useState<string>('');
+    const [didShare, setDidShare] = useState(false);
+    const [showDoneButton, setShowDoneButton] = useState(false);
+
+    const accountName = (abbreviateEnsForDisplay(accountENS, 10) ||
+      formatAddress(accountAddress, 4, 5)) as string;
+
+    useEffect(() => {
+      setAnimationKey(prevKey => prevKey + 1);
+    }, []);
+
+    useImperativeHandle(ref, () => ({
+      isCalculationComplete,
+      setIsCalculationComplete,
+      didConfirmOwnership,
+      setDidConfirmOwnership,
+      showSignInButton,
+      showShareButtons,
+      tweetIntent,
+      didShare,
+      setDidShare,
+      showDoneButton,
+    }));
+
+    usePointsTweetIntentQuery(
+      {
+        id: POINTS_TWEET_INTENT_ID,
+      },
+      {
+        enabled: animationPhase === 2,
+        onSuccess: data => {
+          if (!data.pointsTweetIntent) {
             return;
           }
 
-          // do a string replace to replace {POINTS_VALUE} with their points value
-          let text = tweetIntent.text?.replace(
-            '{POINTS_VALUE}',
-            pointsValue?.toString()
-          );
-          if (Number(metamaskSwaps?.earnings.total) > 0) {
-            text += `, including ${metamaskSwaps?.earnings.total} points because I switched from Metamask`;
+          try {
+            const tweetIntent = data.pointsTweetIntent;
+            const pointsValue =
+              pointsProfile?.onboardPoints?.user.onboarding.earnings.total;
+            if (!pointsValue) {
+              // TODO: Fallback to some default msg
+              return;
+            }
+
+            // do a string replace to replace {POINTS_VALUE} with their points value
+            let text = tweetIntent.text?.replace(
+              '{POINTS_VALUE}',
+              pointsValue?.toString()
+            );
+            if (Number(metamaskSwaps?.earnings.total) > 0) {
+              text += `, including ${metamaskSwaps?.earnings.total} points because I switched from Metamask`;
+            }
+
+            // build the tweet intent url
+            let intent = `https://twitter.com/intent/tweet?text=${encodeURIComponent(
+              text ?? ''
+            )}`;
+            if (tweetIntent.url) {
+              intent += `&url=${encodeURIComponent(tweetIntent.url)}`;
+            }
+
+            if (tweetIntent.via) {
+              intent += `&via=${encodeURIComponent(tweetIntent.via)}`;
+            }
+
+            console.log({ intent });
+            setTweetIntent(intent);
+          } catch (e) {
+            console.log(e);
+            // TODO: add logging
           }
+        },
+      }
+    );
 
-          // build the tweet intent url
-          let intent = `https://twitter.com/intent/tweet?text=${encodeURIComponent(
-            text ?? ''
-          )}`;
-          if (tweetIntent.url) {
-            intent += `&url=${encodeURIComponent(tweetIntent.url)}`;
-          }
+    const onboardingData = pointsProfile?.onboardPoints?.user?.onboarding;
+    const rainbowSwaps = onboardingData?.categories?.find(
+      c => c.type === 'rainbow-swaps'
+    );
+    const metamaskSwaps = onboardingData?.categories?.find(
+      c => c.type === 'metamask-swaps'
+    );
+    const rainbowBridges = onboardingData?.categories?.find(
+      c => c.type === 'metamask-swaps'
+    );
+    const nftCollections = onboardingData?.categories?.find(
+      c => c.type === 'nft-collections'
+    );
+    const historicBalance = onboardingData?.categories?.find(
+      c => c.type === 'historic-balance'
+    );
+    const bonus = onboardingData?.categories?.find(c => c.type === 'bonus');
 
-          if (tweetIntent.via) {
-            intent += `&via=${encodeURIComponent(tweetIntent.via)}`;
-          }
+    const hasRetroactivePoints =
+      rainbowSwaps?.earnings?.total ||
+      metamaskSwaps?.earnings?.total ||
+      rainbowBridges?.earnings?.total ||
+      nftCollections?.earnings?.total ||
+      historicBalance?.earnings?.total;
 
-          console.log({ intent });
-          setTweetIntent(intent);
-        } catch (e) {
-          console.log(e);
-          // TODO: add logging
-        }
-      },
-    }
-  );
+    console.log({ animationKey, animationPhase });
 
-  const onboardingData = pointsProfile?.onboardPoints?.user?.onboarding;
-  const rainbowSwaps = onboardingData?.categories?.find(
-    c => c.type === 'rainbow-swaps'
-  );
-  const metamaskSwaps = onboardingData?.categories?.find(
-    c => c.type === 'metamask-swaps'
-  );
-  const rainbowBridges = onboardingData?.categories?.find(
-    c => c.type === 'metamask-swaps'
-  );
-  const nftCollections = onboardingData?.categories?.find(
-    c => c.type === 'nft-collections'
-  );
-  const historicBalance = onboardingData?.categories?.find(
-    c => c.type === 'historic-balance'
-  );
-  const bonus = onboardingData?.categories?.find(c => c.type === 'bonus');
-
-  const hasRetroactivePoints =
-    rainbowSwaps?.earnings?.total ||
-    metamaskSwaps?.earnings?.total ||
-    rainbowBridges?.earnings?.total ||
-    nftCollections?.earnings?.total ||
-    historicBalance?.earnings?.total;
-
-  return (
-    <TypingAnimation key={animationKey}>
-      {animationPhase === 0 && (
-        <>
-          <Paragraph>
-            <Line>
-              <AnimatedText
-                delayStart={500}
-                color={textColors.gray}
-                textContent={`${i18n.t(i18n.l.points.console.account)}:`}
-                weight="normal"
-              />
-              <AnimatedText
-                color={textColors.account}
-                enableHapticTyping
-                delayStart={300}
-                textContent={accountName}
-              />
-            </Line>
-            <AnimatedText
-              color={textColors.gray}
-              delayStart={500}
-              textContent={`> ${i18n.t(i18n.l.points.console.auth_required)}`}
-              weight="normal"
-            />
-            <AnimatedText
-              color={textColors.gray}
-              onComplete={() => setShowSignInButton(true)}
-              textContent={`> ${i18n.t(
-                i18n.l.points.console.sign_in_with_wallet
-              )}`}
-              weight="normal"
-            />
-            <AnimatedText
-              color={textColors.green}
-              enableHapticTyping
-              startWhenTrue={didConfirmOwnership}
-              textContent={`> ${i18n.t(i18n.l.points.console.access_granted)}`}
-            />
-          </Paragraph>
-          <Paragraph leftIndent={2}>
-            <AnimatedText
-              delayStart={1000}
-              rainbowText
-              textContent={rainbowText.row1}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row2}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row3}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row4}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row5}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row6}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row7}
-              typingSpeed={10}
-            />
-            <AnimatedText
-              rainbowText
-              textContent={rainbowText.row8}
-              typingSpeed={10}
-            />
-            <Line leftIndent={7}>
-              <AnimatedText
-                color={textColors.white}
-                enableHapticTyping
-                onComplete={() => {
-                  const beginNextPhase = setTimeout(() => {
-                    setAnimationKey(prevKey => prevKey + 1);
-                    setAnimationPhase(1);
-                  }, 2500);
-                  return () => clearTimeout(beginNextPhase);
-                }}
-                textContent={rainbowText.row9}
-                typingSpeed={100}
-              />
-            </Line>
-          </Paragraph>
-        </>
-      )}
-      {animationPhase === 1 &&
-        (hasRetroactivePoints ? (
-          <Stack separator={<LineBreak lines={3} />}>
+    return (
+      <TypingAnimation key={animationKey}>
+        {animationPhase === 0 && (
+          <>
             <Paragraph>
               <Line>
                 <AnimatedText
                   delayStart={500}
                   color={textColors.gray}
-                  skipAnimation
                   textContent={`${i18n.t(i18n.l.points.console.account)}:`}
                   weight="normal"
                 />
                 <AnimatedText
                   color={textColors.account}
-                  skipAnimation
+                  enableHapticTyping
+                  delayStart={300}
                   textContent={accountName}
                 />
               </Line>
-              <Line gap={0}>
+              <AnimatedText
+                color={textColors.gray}
+                delayStart={500}
+                textContent={`> ${i18n.t(i18n.l.points.console.auth_required)}`}
+                weight="normal"
+              />
+              <AnimatedText
+                color={textColors.gray}
+                onComplete={() => setShowSignInButton(true)}
+                textContent={`> ${i18n.t(
+                  i18n.l.points.console.sign_in_with_wallet
+                )}`}
+                weight="normal"
+              />
+              <AnimatedText
+                color={textColors.green}
+                enableHapticTyping
+                startWhenTrue={didConfirmOwnership}
+                textContent={`> ${i18n.t(
+                  i18n.l.points.console.access_granted
+                )}`}
+              />
+            </Paragraph>
+            <Paragraph leftIndent={2}>
+              <AnimatedText
+                delayStart={1000}
+                rainbowText
+                textContent={rainbowText.row1}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row2}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row3}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row4}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row5}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row6}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row7}
+                typingSpeed={10}
+              />
+              <AnimatedText
+                rainbowText
+                textContent={rainbowText.row8}
+                typingSpeed={10}
+              />
+              <Line leftIndent={7}>
+                <AnimatedText
+                  color={textColors.white}
+                  enableHapticTyping
+                  onComplete={() => {
+                    const beginNextPhase = setTimeout(() => {
+                      setAnimationKey(prevKey => prevKey + 1);
+                      setAnimationPhase(1);
+                    }, 2500);
+                    return () => clearTimeout(beginNextPhase);
+                  }}
+                  textContent={rainbowText.row9}
+                  typingSpeed={100}
+                />
+              </Line>
+            </Paragraph>
+          </>
+        )}
+        {animationPhase === 1 &&
+          (hasRetroactivePoints ? (
+            <Stack separator={<LineBreak lines={3} />}>
+              <Paragraph>
+                <Line>
+                  <AnimatedText
+                    delayStart={500}
+                    color={textColors.gray}
+                    skipAnimation
+                    textContent={`${i18n.t(i18n.l.points.console.account)}:`}
+                    weight="normal"
+                  />
+                  <AnimatedText
+                    color={textColors.account}
+                    skipAnimation
+                    textContent={accountName}
+                  />
+                </Line>
+                <Line gap={0}>
+                  <AnimatedText
+                    color={textColors.gray}
+                    delayStart={1000}
+                    textContent={`> ${i18n.t(
+                      i18n.l.points.console.calculating_points
+                    )}`}
+                    weight="normal"
+                  />
+                  <AnimatedText
+                    color={textColors.gray}
+                    repeat={!isCalculationComplete}
+                    textContent="..."
+                    typingSpeed={500}
+                    weight="normal"
+                  />
+                </Line>
+              </Paragraph>
+              <Stack separator={<LineBreak lines={2} />}>
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={rainbowColors.blue}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.rainbow_swaps
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={rainbowColors.blue}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textAlign="right"
+                    textContent={`$${abbreviateNumber(
+                      (rainbowSwaps?.data?.usd_amount ?? 0) +
+                        (rainbowBridges?.data?.usd_amount ?? 0)
+                    )}`}
+                    typingSpeed={100}
+                  />
+                </Line>
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={rainbowColors.green}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.rainbow_nfts_owned
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={rainbowColors.green}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textAlign="right"
+                    textContent={`${nftCollections?.data?.owned_collections} of ${nftCollections?.data?.total_collections}`}
+                    typingSpeed={100}
+                  />
+                </Line>
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={rainbowColors.yellow}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.wallet_balance
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={rainbowColors.yellow}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textAlign="right"
+                    textContent={`$${abbreviateNumber(
+                      historicBalance?.data?.usd_amount ?? 0
+                    )}`}
+                    typingSpeed={100}
+                  />
+                </Line>
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={rainbowColors.red}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.metamask_swaps
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={rainbowColors.red}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textAlign="right"
+                    textContent={`$${abbreviateNumber(
+                      metamaskSwaps?.data?.usd_amount ?? 0
+                    )}`}
+                    typingSpeed={100}
+                  />
+                </Line>
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={rainbowColors.purple}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.bonus_points
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={rainbowColors.purple}
+                    delayStart={1000}
+                    enableHapticTyping
+                    onComplete={() => {
+                      setIsCalculationComplete(true);
+                    }}
+                    textAlign="right"
+                    textContent={`+ ${bonus?.earnings?.total}`}
+                    typingSpeed={100}
+                  />
+                </Line>
+              </Stack>
+              <Paragraph gap={30}>
                 <AnimatedText
                   color={textColors.gray}
                   delayStart={1000}
                   textContent={`> ${i18n.t(
-                    i18n.l.points.console.calculating_points
+                    i18n.l.points.console.calculation_complete
                   )}`}
                   weight="normal"
                 />
+                <Line alignHorizontal="justify">
+                  <AnimatedText
+                    color={textColors.white}
+                    delayStart={1000}
+                    enableHapticTyping
+                    textContent={`${i18n.t(
+                      i18n.l.points.console.points_earned
+                    )}:`}
+                  />
+                  <AnimatedText
+                    color={textColors.white}
+                    delayStart={1000}
+                    enableHapticTyping
+                    hapticType="impactHeavy"
+                    textAlign="right"
+                    textContent={(
+                      onboardingData?.earnings?.total ?? 0
+                    ).toLocaleString('en-US')}
+                    typingSpeed={100}
+                  />
+                </Line>
+              </Paragraph>
+            </Stack>
+          ) : (
+            <Stack separator={<LineBreak lines={2} />}>
+              <Paragraph>
+                <Line>
+                  <AnimatedText
+                    delayStart={500}
+                    color={textColors.gray}
+                    skipAnimation
+                    textContent={`${i18n.t(i18n.l.points.console.account)}:`}
+                    weight="normal"
+                  />
+                  <AnimatedText
+                    color={textColors.account}
+                    skipAnimation
+                    textContent={accountName}
+                  />
+                </Line>
                 <AnimatedText
                   color={textColors.gray}
-                  repeat={!isCalculationComplete}
-                  textContent="..."
-                  typingSpeed={500}
-                  weight="normal"
-                />
-              </Line>
-            </Paragraph>
-            <Stack separator={<LineBreak lines={2} />}>
-              <Line alignHorizontal="justify">
-                <AnimatedText
-                  color={rainbowColors.blue}
-                  enableHapticTyping
-                  textContent={`${i18n.t(
-                    i18n.l.points.console.rainbow_swaps
-                  )}:`}
-                />
-                <AnimatedText
-                  color={rainbowColors.blue}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textAlign="right"
-                  textContent={`$${abbreviateNumber(
-                    (rainbowSwaps?.data?.usd_amount ?? 0) +
-                      (rainbowBridges?.data?.usd_amount ?? 0)
+                  textContent={`> ${i18n.t(
+                    i18n.l.points.console.calculating_points
                   )}`}
-                  typingSpeed={100}
                 />
-              </Line>
-              <Line alignHorizontal="justify">
-                <AnimatedText
-                  color={rainbowColors.green}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textContent={`${i18n.t(
-                    i18n.l.points.console.rainbow_nfts_owned
-                  )}:`}
-                />
-                <AnimatedText
-                  color={rainbowColors.green}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textAlign="right"
-                  textContent={`${nftCollections?.data?.owned_collections} of ${nftCollections?.data?.total_collections}`}
-                  typingSpeed={100}
-                />
-              </Line>
-              <Line alignHorizontal="justify">
-                <AnimatedText
-                  color={rainbowColors.yellow}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textContent={`${i18n.t(
-                    i18n.l.points.console.wallet_balance
-                  )}:`}
-                />
-                <AnimatedText
-                  color={rainbowColors.yellow}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textAlign="right"
-                  textContent={`$${abbreviateNumber(
-                    historicBalance?.data?.usd_amount ?? 0
-                  )}`}
-                  typingSpeed={100}
-                />
-              </Line>
-              <Line alignHorizontal="justify">
-                <AnimatedText
-                  color={rainbowColors.red}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textContent={`${i18n.t(
-                    i18n.l.points.console.metamask_swaps
-                  )}:`}
-                />
-                <AnimatedText
-                  color={rainbowColors.red}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textAlign="right"
-                  textContent={`$${abbreviateNumber(
-                    metamaskSwaps?.data?.usd_amount ?? 0
-                  )}`}
-                  typingSpeed={100}
-                />
-              </Line>
+              </Paragraph>
               <Line alignHorizontal="justify">
                 <AnimatedText
                   color={rainbowColors.purple}
@@ -657,49 +630,27 @@ const ClaimRetroactivePointsFlow = ({
                   color={rainbowColors.purple}
                   delayStart={1000}
                   enableHapticTyping
-                  onComplete={() => {
-                    setAnimationKey(prevKey => prevKey + 1);
-                    setIsCalculationComplete(true);
-                  }}
                   textAlign="right"
                   textContent={`+ ${bonus?.earnings?.total}`}
                   typingSpeed={100}
                 />
               </Line>
-            </Stack>
-            <Paragraph gap={30}>
               <AnimatedText
                 color={textColors.gray}
                 delayStart={1000}
-                textContent={`> ${i18n.t(
-                  i18n.l.points.console.calculation_complete
-                )}`}
+                onComplete={() => {
+                  setIsCalculationComplete(true);
+                }}
                 weight="normal"
+                multiline
+                textContent={i18n.t(
+                  i18n.l.points.console.claim_bonus_paragraph
+                )}
               />
-              <Line alignHorizontal="justify">
-                <AnimatedText
-                  color={textColors.white}
-                  delayStart={1000}
-                  enableHapticTyping
-                  textContent={`${i18n.t(
-                    i18n.l.points.console.points_earned
-                  )}:`}
-                />
-                <AnimatedText
-                  color={textColors.white}
-                  delayStart={1000}
-                  enableHapticTyping
-                  hapticType="impactHeavy"
-                  textAlign="right"
-                  textContent={(
-                    onboardingData?.earnings?.total ?? 0
-                  ).toLocaleString('en-US')}
-                  typingSpeed={100}
-                />
-              </Line>
-            </Paragraph>
-          </Stack>
-        ) : (
+            </Stack>
+          ))}
+
+        {animationPhase === 2 && (
           <Stack separator={<LineBreak lines={2} />}>
             <Paragraph>
               <Line>
@@ -719,533 +670,112 @@ const ClaimRetroactivePointsFlow = ({
               <AnimatedText
                 color={textColors.gray}
                 textContent={`> ${i18n.t(
-                  i18n.l.points.console.calculating_points
+                  i18n.l.points.console.referral_link_is_ready
                 )}`}
               />
             </Paragraph>
-            <Line alignHorizontal="justify">
-              <AnimatedText
-                color={rainbowColors.purple}
-                delayStart={1000}
-                enableHapticTyping
-                textContent={`${i18n.t(i18n.l.points.console.bonus_points)}:`}
-              />
-              <AnimatedText
-                color={rainbowColors.purple}
-                delayStart={1000}
-                enableHapticTyping
-                textAlign="right"
-                textContent={`+ ${bonus?.earnings?.total}`}
-                typingSpeed={100}
-              />
-            </Line>
             <AnimatedText
-              color={textColors.gray}
+              color={rainbowColors.yellow}
+              delayStart={1000}
+              weight="normal"
+              multiline
+              textContent={`${i18n.t(
+                i18n.l.points.console.referral_link_bonus_text
+              )}:`}
+            />
+            <AnimatedText
+              color={rainbowColors.yellow}
               delayStart={1000}
               onComplete={() => {
                 setAnimationKey(prevKey => prevKey + 1);
-                setIsCalculationComplete(true);
+                setShowShareButtons(true);
               }}
               weight="normal"
               multiline
-              textContent={i18n.t(i18n.l.points.console.claim_bonus_paragraph)}
+              textContent={i18n.t(
+                i18n.l.points.console.referral_link_bonus_text_extended
+              )}
             />
           </Stack>
-        ))}
+        )}
 
-      {animationPhase === 2 && (
-        <Stack separator={<LineBreak lines={2} />}>
-          <Paragraph>
-            <Line>
+        {animationPhase === 3 && (
+          <Stack separator={<LineBreak lines={2} />}>
+            <Paragraph>
+              <Line>
+                <AnimatedText
+                  delayStart={500}
+                  color={textColors.gray}
+                  skipAnimation
+                  textContent={`${i18n.t(i18n.l.points.console.account)}:`}
+                  weight="normal"
+                />
+                <AnimatedText
+                  color={textColors.account}
+                  skipAnimation
+                  textContent={accountName}
+                />
+              </Line>
               <AnimatedText
-                delayStart={500}
-                color={textColors.gray}
-                skipAnimation
-                textContent={`${i18n.t(i18n.l.points.console.account)}:`}
-                weight="normal"
+                color={textColors.green}
+                textContent={`> ${i18n.t(
+                  i18n.l.points.console.referral_link_is_ready
+                )}`}
               />
-              <AnimatedText
-                color={textColors.account}
-                skipAnimation
-                textContent={accountName}
-              />
-            </Line>
-            <AnimatedText
-              color={textColors.gray}
-              textContent={`> ${i18n.t(
-                i18n.l.points.console.referral_link_is_ready
-              )}`}
-            />
-          </Paragraph>
-          <AnimatedText
-            color={rainbowColors.yellow}
-            delayStart={1000}
-            weight="normal"
-            multiline
-            textContent={`${i18n.t(
-              i18n.l.points.console.referral_link_bonus_text
-            )}:`}
-          />
-          <AnimatedText
-            color={rainbowColors.yellow}
-            delayStart={1000}
-            onComplete={() => {
-              setAnimationKey(prevKey => prevKey + 1);
-              setShowShareButtons(true);
-            }}
-            weight="normal"
-            multiline
-            textContent={i18n.t(
-              i18n.l.points.console.referral_link_bonus_text_extended
+            </Paragraph>
+            {didShare && (
+              <Line alignHorizontal="justify">
+                <AnimatedText
+                  color={rainbowColors.yellow}
+                  delayStart={1000}
+                  enableHapticTyping
+                  textContent={`${i18n.t(i18n.l.points.console.share_bonus)}:`}
+                />
+                <AnimatedText
+                  color={rainbowColors.yellow}
+                  delayStart={1000}
+                  enableHapticTyping
+                  textAlign="right"
+                  textContent={`+ 250`} // TODO: Where does this amount come from?
+                />
+              </Line>
             )}
-          />
-        </Stack>
-      )}
-
-      {animationPhase === 3 && (
-        <Stack separator={<LineBreak lines={2} />}>
-          <Paragraph>
-            <Line>
+            <Paragraph>
               <AnimatedText
                 delayStart={500}
                 color={textColors.gray}
-                skipAnimation
-                textContent={`${i18n.t(i18n.l.points.console.account)}:`}
+                textContent={`${i18n.t(
+                  i18n.l.points.console.share_bonus_paragraph_one
+                )}:`}
                 weight="normal"
               />
               <AnimatedText
-                color={textColors.account}
-                skipAnimation
-                textContent={accountName}
-              />
-            </Line>
-            <AnimatedText
-              color={textColors.green}
-              textContent={`> ${i18n.t(
-                i18n.l.points.console.referral_link_is_ready
-              )}`}
-            />
-          </Paragraph>
-          {didShare && (
-            <Line alignHorizontal="justify">
-              <AnimatedText
-                color={rainbowColors.yellow}
-                delayStart={1000}
-                enableHapticTyping
-                textContent={`${i18n.t(i18n.l.points.console.share_bonus)}:`}
+                color={textColors.gray}
+                delayStart={500}
+                textContent={`> ${i18n.t(
+                  i18n.l.points.console.share_bonus_paragraph_two
+                )}`}
+                weight="normal"
               />
               <AnimatedText
-                color={rainbowColors.yellow}
-                delayStart={1000}
-                enableHapticTyping
-                textAlign="right"
-                textContent={`+ 250`} // TODO: Where does this amount come from?
+                color={textColors.gray}
+                onComplete={() => setShowDoneButton(true)}
+                textContent={`> ${i18n.t(
+                  i18n.l.points.console.share_bonus_paragraph_three
+                )}`}
+                weight="normal"
               />
-            </Line>
-          )}
-          <Paragraph>
-            <AnimatedText
-              delayStart={500}
-              color={textColors.gray}
-              textContent={`${i18n.t(
-                i18n.l.points.console.share_bonus_paragraph_one
-              )}:`}
-              weight="normal"
-            />
-            <AnimatedText
-              color={textColors.gray}
-              delayStart={500}
-              textContent={`> ${i18n.t(
-                i18n.l.points.console.share_bonus_paragraph_two
-              )}`}
-              weight="normal"
-            />
-            <AnimatedText
-              color={textColors.gray}
-              onComplete={() => setShowDoneButton(true)}
-              textContent={`> ${i18n.t(
-                i18n.l.points.console.share_bonus_paragraph_three
-              )}`}
-              weight="normal"
-            />
-          </Paragraph>
-        </Stack>
-      )}
-    </TypingAnimation>
-  );
-};
-
-const AnimationContext = createContext({
-  currentSequenceIndex: 0,
-  getNextAnimationIndex: () => {
-    return;
-  },
-  incrementSequence: () => {
-    return;
-  },
-});
-
-export const useAnimationContext = () => useContext(AnimationContext);
-
-export const TypingAnimation = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  const [currentSequenceIndex, setCurrentSequenceIndex] = useState(0);
-  const animationIndexRef = useRef(0);
-
-  const getNextAnimationIndex = useCallback(() => {
-    const currentIndex = animationIndexRef.current;
-    animationIndexRef.current += 1;
-    return currentIndex;
-  }, []);
-
-  const incrementSequence = useCallback(() => {
-    setCurrentSequenceIndex(prevIndex => prevIndex + 1);
-  }, []);
-
-  useEffect(() => {
-    setCurrentSequenceIndex(0);
-    animationIndexRef.current = 0;
-  }, []);
-
-  return (
-    <AnimationContext.Provider
-      value={{ currentSequenceIndex, getNextAnimationIndex, incrementSequence }}
-    >
-      {children}
-    </AnimationContext.Provider>
-  );
-};
-
-type AnimatedTextProps = {
-  color?: { text: string; shadow: string };
-  delayStart?: number;
-  disableShadow?: boolean;
-  enableHapticTyping?: boolean;
-  hapticType?: HapticFeedbackType;
-  multiline?: boolean;
-  onComplete?: () => void;
-  opacity?: number;
-  rainbowText?: boolean;
-  repeat?: boolean;
-  shadowOpacity?: number;
-  skipAnimation?: boolean;
-  startWhenTrue?: boolean;
-  textAlign?: 'center' | 'left' | 'right';
-  textContent: string;
-  typingSpeed?: number;
-  weight?: 'bold' | 'normal';
-} & ({ color: { text: string; shadow: string } } | { rainbowText: boolean });
-
-const AnimatedText = ({
-  color,
-  delayStart,
-  disableShadow,
-  enableHapticTyping,
-  hapticType = 'selection',
-  multiline,
-  onComplete,
-  opacity,
-  rainbowText,
-  repeat,
-  shadowOpacity,
-  skipAnimation,
-  startWhenTrue,
-  textAlign,
-  textContent,
-  typingSpeed = 20,
-  weight = 'bold',
-}: AnimatedTextProps) => {
-  const { colors } = useTheme();
-  const {
-    currentSequenceIndex,
-    getNextAnimationIndex,
-    incrementSequence,
-  } = useAnimationContext();
-  const index = useRef(getNextAnimationIndex()).current;
-  const displayedCharacters = useSharedValue(
-    skipAnimation ? textContent.length : 0
-  );
-  const [displayedText, setDisplayedText] = useState(
-    skipAnimation ? textContent : ''
-  );
-
-  const rainbowTextColors = useMemo(
-    () => (rainbowText ? generateRainbowColors(textContent) : undefined),
-    [rainbowText, textContent]
-  );
-
-  const getRainbowTextStyle = useCallback(
-    (i: number) => ({
-      color: rainbowTextColors?.[i]?.text,
-      opacity,
-      textAlign,
-      textShadowColor: disableShadow
-        ? 'transparent'
-        : shadowOpacity && rainbowTextColors?.[i]?.shadow
-        ? colors.alpha(rainbowTextColors?.[i]?.shadow, shadowOpacity)
-        : rainbowTextColors?.[i]?.shadow,
-    }),
-    [
-      colors,
-      disableShadow,
-      opacity,
-      rainbowTextColors,
-      shadowOpacity,
-      textAlign,
-    ]
-  );
-
-  const textStyle = useMemo(
-    () => ({
-      color: rainbowText ? undefined : color?.text,
-      fontWeight: weight,
-      opacity,
-      textAlign,
-      textShadowColor: disableShadow
-        ? 'transparent'
-        : rainbowText
-        ? undefined
-        : shadowOpacity && color?.shadow
-        ? colors.alpha(color?.shadow, shadowOpacity)
-        : color?.shadow,
-    }),
-    [
-      color,
-      colors,
-      disableShadow,
-      opacity,
-      rainbowText,
-      shadowOpacity,
-      textAlign,
-      weight,
-    ]
-  );
-
-  const animationConfig = useMemo(
-    () => ({
-      duration: textContent.length * typingSpeed,
-      easing: Easing.linear,
-    }),
-    [textContent, typingSpeed]
-  );
-
-  const onAnimationComplete = useCallback(
-    (isFinished?: boolean) => {
-      'worklet';
-      if (isFinished) {
-        if (onComplete) {
-          runOnJS(onComplete)();
-        }
-        runOnJS(incrementSequence)();
-
-        if (repeat) {
-          displayedCharacters.value = withRepeat(
-            withSequence(
-              withTiming(textContent.length, { duration: typingSpeed }),
-              withTiming(0, { duration: 0 }),
-              withTiming(0, { duration: typingSpeed }),
-              withTiming(textContent.length, animationConfig)
-            ),
-            -1,
-            false
-          );
-        }
-      }
-    },
-    [
-      animationConfig,
-      displayedCharacters,
-      incrementSequence,
-      onComplete,
-      repeat,
-      textContent.length,
-      typingSpeed,
-    ]
-  );
-
-  useAnimatedReaction(
-    () => ({ displayedValue: displayedCharacters.value, repeat }),
-    (current, previous) => {
-      if (
-        !previous?.displayedValue ||
-        Math.round(current.displayedValue) !==
-          Math.round(previous?.displayedValue)
-      ) {
-        const newText =
-          textContent.slice(0, Math.round(current.displayedValue)) || ' ';
-
-        if (current.repeat === false && newText === textContent) {
-          runOnJS(setDisplayedText)(newText);
-          cancelAnimation(displayedCharacters);
-          displayedCharacters.value = textContent.length;
-          return;
-        }
-
-        runOnJS(setDisplayedText)(newText);
-        if (
-          enableHapticTyping &&
-          Math.round(current.displayedValue) &&
-          newText[newText.length - 1] !== ' '
-        ) {
-          runOnJS(triggerHapticFeedback)(hapticType);
-        }
-      }
-    }
-  );
-
-  useEffect(() => {
-    if (
-      index !== undefined &&
-      currentSequenceIndex === index &&
-      (startWhenTrue === undefined || startWhenTrue)
-    ) {
-      if (!skipAnimation) {
-        const timer = setTimeout(() => {
-          displayedCharacters.value = 0;
-          displayedCharacters.value = withTiming(
-            textContent.length,
-            animationConfig,
-            onAnimationComplete
-          );
-        }, delayStart || 0);
-
-        return () => {
-          clearTimeout(timer);
-        };
-      } else {
-        onComplete?.();
-        incrementSequence();
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentSequenceIndex, index, startWhenTrue, textContent]);
-
-  return (
-    <Bleed space="16px">
-      <RNText
-        style={[styles.text, textStyle, multiline ? { lineHeight: 20.25 } : {}]}
-      >
-        {rainbowText
-          ? displayedText.split('').map((char, i) => (
-              <RNText key={i} style={getRainbowTextStyle(i)}>
-                {char}
-              </RNText>
-            ))
-          : displayedText}
-      </RNText>
-    </Bleed>
-  );
-};
-
-const Line = ({
-  alignHorizontal,
-  children,
-  gap = 10,
-  leftIndent = 0,
-}: {
-  alignHorizontal?: 'center' | 'justify' | 'left' | 'right';
-  children: React.ReactNode;
-  gap?: number;
-  leftIndent?: number;
-}) => {
-  return (
-    <View
-      style={{
-        flexDirection: 'row',
-        flexWrap: 'wrap',
-        gap,
-        justifyContent: alignHorizontal
-          ? alignHorizontalToFlexAlign[alignHorizontal]
-          : undefined,
-        paddingLeft: leftIndent * CHARACTER_WIDTH,
-      }}
-    >
-      {children}
-    </View>
-  );
-};
-
-const LineBreak = ({ lines = 1 }: { lines?: number }) => {
-  return <View style={{ height: lines * 15 }} />;
-};
-
-const Paragraph = ({
-  children,
-  gap = 15,
-  leftIndent = 0,
-}: {
-  children: React.ReactNode;
-  gap?: number;
-  leftIndent?: number;
-}) => {
-  return (
-    <View
-      style={{
-        flexWrap: 'wrap',
-        gap,
-        paddingLeft: leftIndent * CHARACTER_WIDTH,
-      }}
-    >
-      {children}
-    </View>
-  );
-};
-
-const generateRainbowColors = (
-  text: string
-): Array<{ text: string; shadow: string }> | undefined => {
-  let colorIndex = 0;
-  let repeatCount = 0;
-  const colorKeys: string[] = Object.keys(rainbowColors);
-  const colors: Array<{ text: string; shadow: string }> = [];
-  const repeatLength: number = Math.floor(text.length / (colorKeys.length * 2));
-
-  text.split('').forEach(() => {
-    if (repeatCount >= repeatLength + Math.round(Math.random())) {
-      repeatCount = 0;
-      colorIndex = (colorIndex + 1) % colorKeys.length;
-    }
-    colors.push(
-      rainbowColors[colorKeys[colorIndex] as keyof typeof rainbowColors]
+            </Paragraph>
+          </Stack>
+        )}
+      </TypingAnimation>
     );
-    repeatCount += 1;
-  });
+  }
+);
 
-  return colors;
-};
-
-const triggerHapticFeedback = (hapticType: HapticFeedbackType) =>
-  ReactNativeHapticFeedback.trigger(hapticType);
+ClaimRetroactivePointsFlow.displayName = 'ClaimRetroactivePointsFlow';
 
 const styles = StyleSheet.create({
-  neonButtonWrapper: {
-    alignSelf: 'center',
-  },
-  neonButton: {
-    alignContent: 'center',
-    backgroundColor: '#191A1C',
-    borderCurve: 'continuous',
-    borderRadius: 12,
-    borderWidth: 1.5,
-    height: 48,
-    justifyContent: 'center',
-    shadowOffset: { width: 0, height: 13 },
-    shadowOpacity: 0.2,
-    shadowRadius: 26,
-  },
-  neonButtonFill: {
-    marginLeft: -((1 / 3) * 2),
-    marginTop: -((1 / 3) * 2),
-  },
-  neonButtonText: {
-    margin: -16,
-    padding: 16,
-    textShadowOffset: { width: 0, height: 0 },
-    textShadowRadius: 12,
-  },
   sheet: {
     backgroundColor: '#191A1C',
     borderColor: 'rgba(245, 248, 255, 0.06)',
@@ -1259,41 +789,4 @@ const styles = StyleSheet.create({
     width: '100%',
     zIndex: -1,
   },
-  text: {
-    fontFamily: fonts.family.SFMono,
-    fontSize: 15,
-    fontWeight: 'bold',
-    lineHeight: 11,
-    overflow: 'visible',
-    padding: 16,
-    textShadowOffset: { width: 0, height: 0 },
-    textShadowRadius: 12,
-  },
 });
-
-const rainbowColors = {
-  blue: { text: '#31BCC4', shadow: 'rgba(49, 188, 196, 0.8)' },
-  green: { text: '#57EA5F', shadow: 'rgba(87, 234, 95, 0.8)' },
-  yellow: { text: '#F0D83F', shadow: 'rgba(240, 216, 63, 0.8)' },
-  red: { text: '#DF5337', shadow: 'rgba(223, 83, 55, 0.8)' },
-  purple: { text: '#B756A7', shadow: 'rgba(183, 86, 167, 0.8)' },
-};
-
-const textColors = {
-  account: { text: '#FEC101', shadow: 'rgba(254, 193, 1, 0.8)' },
-  gray: { text: '#94969B', shadow: 'rgba(148, 150, 155, 0.8)' },
-  green: { text: '#3ECF5B', shadow: 'rgba(62, 207, 91, 0.8)' },
-  white: { text: '#FFFFFF', shadow: 'rgba(255, 255, 255, 0.8)' },
-};
-
-const rainbowText = {
-  row1: '\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row2: ' \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row3: '  \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row4: '   \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row5: '    \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row6: '     \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row7: '      \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row8: '       \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\',
-  row9: ' WELCOME TO POINTS ',
-};
