+++ src/screens/points/components/LineBreak.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export const LineBreak = ({ lines = 1 }: { lines?: number }) => {
+  return <View style={{ height: lines * 15 }} />;
+};
+
+export default LineBreak;
