+++ src/screens/points/content/PointsContent.tsx
@@ -33,7 +33,7 @@ import {
   ButtonPressAnimation,
   ShimmerAnimation,
 } from '@/components/animations';
-import { TAB_BAR_HEIGHT } from '@/navigation/SwipeNavigator';
+import { getHeaderHeight } from '@/navigation/SwipeNavigator';
 import { addressCopiedToastAtom } from '@/recoil/addressCopiedToastAtom';
 import { useRecoilState } from 'recoil';
 import * as i18n from '@/languages';
@@ -361,10 +361,10 @@ export default function PointsContent() {
     <Box height="full" background="surfacePrimary" as={Page} flex={1}>
       <ScrollView
         scrollIndicatorInsets={{
-          bottom: TAB_BAR_HEIGHT - safeAreaInsetValues.bottom,
+          bottom: getHeaderHeight() - safeAreaInsetValues.bottom,
         }}
         contentContainerStyle={{
-          paddingBottom: TAB_BAR_HEIGHT + 32,
+          paddingBottom: getHeaderHeight() + 32,
           height: !shouldDisplayError ? undefined : '100%',
         }}
         showsVerticalScrollIndicator={!shouldDisplayError}
